pyproject.toml
[tool.poetry]
name = "reddit-bot"
version = "0.1.0"
description = "A bot that monitors Reddit posts and forwards them to Discord and Telegram"
authors = ["laraib_sidd <laraib.sidd7@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "3.11.8"
discord-webhook = "^1.3.0"
praw = "^7.7.1"
pygithub = "^2.5.0"
pymongo = "^4.7.0"
pynacl = "^1.5.0"
python-telegram-bot = "^21.1.1"
pytz = "^2024.1"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
telegram-bot = "scripts.telegram_bot:main"
discord-bot = "scripts.discord_bot:main"
sync-secrets = "scripts.sync_secrets:main"
test-secret-value = "tests.test_secret_value:main"
run-tests = "scripts.run_tests:main"
live-test = "scripts.live_test:run_main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

